Apply changes from code we received from a retired Bisq dev.
As we don't have a git version to know on which version was based on, it was hard to merge.
I tried to not delete any files from latest snapshot (before Bisq got removed from mempool) and I tried to keep new changes which are likely unreleated to his changes. Though its messy and requires a dev familiar with the used tech stack to clear all that up.